Revise color grading naming conventions
Previous color grades were closely related to computed luminance values. This approach is unnecessarily complex and requires scrutiny for each value update. Instead, use numbers that represent the lightness or darkness of a color: the higher the number, the darker the color. Luminance measurements should still be employed, but only as a means of ensuring colors are perceptually similar.